REFACTOR [server.ts, docker compose, scripts] Refactor Express server.ts into routes and db
