+++ server.ts
@@ -3,26 +3,16 @@ import { CommonEngine } from '@angular/ssr';
 import express from 'express';
 import { fileURLToPath } from 'node:url';
 import { dirname, join, resolve } from 'node:path';
-import { createProxyMiddleware } from 'http-proxy-middleware';
 import { environment } from './src/environments/environment';
 import AppServerModule from './src/main.server';
-import axios from 'axios';
 import dotenv from 'dotenv';
-import { Octokit } from '@octokit/rest';
-import { Pool } from 'pg';
+
+import { githubRouter } from './routes/github';
+import { usersRouter } from './routes/users';
+import { repositoriesRouter } from './routes/repositories';
+import { reportsRouter } from './routes/reports';
 
 dotenv.config();
-
-// Initialize database connection
-const pool = new Pool({
-  host: environment.dbHost,
-  port: parseInt(environment.dbPort),
-  database: environment.dbName,
-  user: environment.dbUser,
-  password: environment.dbPassword,
-  ssl: false
-});
-console.log(`Pool environment:`, environment)
 
 export function app(): express.Express {
   const server = express();
@@ -36,90 +26,11 @@ export function app(): express.Express {
   server.set('views', browserDistFolder);
   server.use(express.json());
 
-  // GitHub OAuth callback endpoint
-  server.post('/api/github/callback', async (req, res) => {
-    const { code } = req.body;
-    try {
-      const response = await axios.post('https://github.com/login/oauth/access_token', {
-        client_id: environment.githubClientId,
-        client_secret: environment.githubClientSecret,
-        code,
-      }, {
-        headers: {
-          Accept: 'application/json'
-        }
-      });
-      res.json(response.data);
-    } catch (error) {
-      console.error('GitHub OAuth error:', error);
-      res.status(500).json({ error: 'Failed to authenticate' });
-    }
-  });
-
-  // GitHub API proxy (now using Octokit)
-  server.get('/api/github/issues/:owner/:repo', async (req, res) => {
-    const { owner, repo } = req.params;
-    const token = req.headers.authorization?.split(' ')[1];
-    
-    try {
-      const octokit = new Octokit({ auth: token });
-      const { data } = await octokit.issues.listForRepo({
-        owner,
-        repo,
-        per_page: token ? 30 : 10,
-        state: 'open'
-      });
-
-      res.json(data);
-    } catch (error: any) {
-      console.error('GitHub API error:', error);
-      res.status(error.status || 500).json({ error: 'Failed to fetch issues' });
-    }
-  });
-
-  // PostgreSQL Database Endpoint
-  server.get('/api/reports/:id', async (req, res) => {
-    const reportId = parseInt(req.params.id);
-    console.log("FETCHING", reportId)
-    if (isNaN(reportId)) {
-      return res.status(400).json({ error: 'Invalid report ID' });
-    }
-  
-    try {
-      const query = `
-        SELECT * from reports WHERE reportID = $1;
-      `;
-      const result = await pool.query(query, [reportId]);
-      console.log("result:", result)
-      if (result.rows.length === 0) {
-        return res.status(404).json({ error: 'Report not found' });
-      }
-  
-      const report = result.rows[0];
-      return res.json({
-        reportid: report.reportid,
-        creatorid: report.creatorid,
-        datecreated: report.datecreated,
-        isopen: report.isopen,
-        repoid: report.repoid,
-        repoadmingithubid: report.repoadmingithubid,
-        reportcontent: report.reportcontent
-      });
-    } catch (err) {
-      console.error('Error fetching report:', err);
-      return res.status(500).json({ error: 'An error occurred while fetching the report' });
-    }
-  });
-
-  // Proxy requests to FastAPI backend
-  server.use('/api', createProxyMiddleware({
-    target: environment.fastApiUrl,
-    changeOrigin: true,
-    pathRewrite: {
-      '^/api/classify_spam': '/classify_spam',
-      '^/api/issues': '/issues'
-    }
-  }));
+  // Use routers
+  server.use('/api/github', githubRouter);
+  server.use('/api/users', usersRouter);
+  server.use('/api/repositories', repositoriesRouter);
+  server.use('/api/reports', reportsRouter);
 
   server.get('**', express.static(browserDistFolder, {
     maxAge: '1y',
@@ -148,7 +59,6 @@ export function app(): express.Express {
 function run(): void {
   const port = 3000;
 
-  // Start up the Node server
   const server = app();
   server.listen(port, () => {
     console.log(`Node Express server listening on http://localhost:${port}`);
