--- start-dev.bat
@@ -1,3 +1,0 @@
-@echo off
-
-ng build && npm run serve:ssr:repo-cleanup
